권한체크

+++ src/main/java/com/koreait/surl_project_11/domain/surl/surl/controller/ApiV1SurlController.java
@@ -76,6 +76,12 @@ public class ApiV1SurlController {
     ) {
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
 
+        Member member = rq.getMember();
+
+        if (!surl.getAuthor().equals(member)) {
+            throw new GlobalException("403-1", "권한이 없어");
+        }
+
         return RsData.of(
                 new SurlGetRespBody(
                         new SurlDto(surl)
@@ -113,6 +119,13 @@ public class ApiV1SurlController {
             @PathVariable long id
     ) {
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
+
+        Member member = rq.getMember();
+
+        if (!surl.getAuthor().equals(member)) {
+            throw new GlobalException("403-1", "권한이 없어");
+        }
+
         surlService.delete(surl);
         return RsData.OK;
     }
@@ -139,6 +152,13 @@ public class ApiV1SurlController {
             @RequestBody @Valid SurlModifyReqBody reqBody
     ) {
         Surl surl = surlService.findById(id).orElseThrow(GlobalException.E404::new);
+
+        Member member = rq.getMember();
+
+        if (!surl.getAuthor().equals(member)) {     // 객체 비교는 equals 사용해야함. id의 타입이 Long 이고, Long은 기본타입이 아니기 때문에 equals를 사용해야함(참조타입)
+//        if (surl.getAuthor().getId() != member.getId()) {     // 이건 단순 값비교할때 사용
+            throw new GlobalException("403-1", "권한이 없어");
+        }
 
         RsData<Surl> modifyRs = surlService.modify(surl, reqBody.body, reqBody.url);
